refactor: :truck: update folder name
